+++ src/main/java/br/com/projeto/contratados/domain/repository/SolicitacaoRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SolicitacaoRepository extends JpaRepository<Solicitacao, Integer> {
-    boolean existsByUsuarioId(Integer id);
+public interface SolicitacaoRepository extends JpaRepository<Solicitacao, Long> {
+    boolean existsByUsuarioId(Long id);
 
-    boolean existsByAnuncioVagaId(Integer id);
+    boolean existsByAnuncioVagaId(Long id);
 }
